# docs/postman_OpenApiV3_response.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Car-Rest-Service API",
    "version": "v1"
  },
  "servers": [
    {
      "url": "http://localhost:8000",
      "description": "Generated server url"
    }
  ],
...
  "paths": {
    "/api/v1/models/{modelId}": {
      "get": {
        "tags": [
          "models"
        ],
        "summary": "Get model by id",
        "description": "Retrieve a specific model by its unique id",
        "operationId": "getModelById",
        "parameters": [
          {
            "name": "modelId",
            "in": "path",
            "description": "Unique identifier of the model",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            },
            "example": 1
          }
        ],
        "responses": {
          "200": {
            "description": "Model retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ModelDto"
                }
              }
            }
          },
          "400": {
            "description": "invalid modelId format"
          },
          "404": {
            "description": "Model not found with provided id"
          }
        }
      },
      "put": {
        "tags": [
          "models"
        ],
        "summary": "Update model by id",
        "description": "Update an existing model. Requires authentification",
        "operationId": "updateModel",
        "parameters": [
          {
            "name": "modelId",
            "in": "path",
            "description": "Unique identifier of the model",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            },
            "example": 1
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ModelDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Model updated successfully"
          },
          "400": {
            "description": "Invalid input data or modelId format"
          },
          "401": {
            "description": "Unauthorized - Invalid or missing JWT token"
          },
          "404": {
            "description": "Model not found with provided id"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      },
      "delete": {
        "tags": [
          "models"
        ],
        "summary": "Delete model by id",
        "description": "Delete a specific model. Requires authentification.",
        "operationId": "deleteModelById",
        "parameters": [
          {
            "name": "modelId",
            "in": "path",
            "description": "Unique identifier of the model",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            },
            "example": 1
          }
        ],
        "responses": {
          "204": {
            "description": "Model deleted successfully"
          },
          "400": {
            "description": "Invalid modelId format"
          },
          "401": {
            "description": "Unauthorized - Invalid or missing JWT token"
          },
          "404": {
            "description": "Model not found with provided id"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
    "/api/v1/manufacturers/{manufacturerName}/models/{modelName}": {
      "get": {
        "tags": [
          "manufacturers"
        ],
        "summary": "Get model by manufacturer and model name",
        "description": "Retrieve a specific model for a manufacturer by model name",
        "operationId": "getModelByManufacturerAndName",
        "parameters": [
          {
            "name": "manufacturerName",
            "in": "path",
            "description": "Name of the manufacturer",
            "required": true,
            "schema": {
              "type": "string"
            },
            "example": "Toyota"
          },
          {
            "name": "modelName",
            "in": "path",
            "description": "Name of the model",
            "required": true,
            "schema": {
              "type": "string"
            },
            "example": "Camry"
          }
        ],
        "responses": {
          "200": {
            "description": "Model retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ModelDto"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input data"
          },
          "404": {
            "description": "Manufacturer or model not found with provided name"
          }
        }
      },
      "put": {
        "tags": [
          "manufacturers"
        ],
        "summary": "Update model by manufacturer and model name",
        "description": "Update an existing model by a manufacturer. Requires authentification.",
        "operationId": "updateModelByManufacturerAndName",
        "parameters": [
          {
            "name": "manufacturerName",
            "in": "path",
            "description": "Name of the manufacturer",
            "required": true,
            "schema": {
              "type": "string"
            },
            "example": "Toyota"
          },
          {
            "name": "modelName",
            "in": "path",
            "description": "Name of the model",
            "required": true,
            "schema": {
              "type": "string"
            },
            "example": "Camry"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ModelDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Model updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ModelDto"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input data or parameter format"
          },
          "401": {
            "description": "Unauthorized - Invalid or missing JWT token"
          },
          "404": {
            "description": "Manufacturer or model not found with provided name"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      },
      "delete": {
        "tags": [
          "manufacturers"
        ],
        "summary": "Delete model by manufacturer and model name",
        "description": "Delete a specific model by a manufacturer. Requires authentification.",
        "operationId": "deleteModelByManufacturerAndName",
        "parameters": [
          {
            "name": "manufacturerName",
            "in": "path",
            "description": "Name of the manufacturer",
            "required": true,
            "schema": {
              "type": "string"
            },
            "example": "Toyota"
          },
          {
            "name": "modelName",
            "in": "path",
            "description": "Name of the model",
            "required": true,
            "schema": {
              "type": "string"
            },
            "example": "Camry"
          }
        ],
        "responses": {
          "204": {
            "description": "Model deleted successfully"
          },
          "400": {
            "description": "Invalid parameter format"
          },
          "401": {
            "description": "Unauthorized - Invalid or missing JWT token"
          },
          "404": {
            "description": "Manufacturer or model not found with provided name"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
    "/api/v1/models": {
      "get": {
        "tags": [
          "models"
        ],
        "summary": "Get all models with pagination",
        "description": "Retrieve a paginated list of all models",
        "operationId": "getAllModels",
        "parameters": [
          {
            "name": "pageable",
            "in": "query",
            "description": "Pagination information",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/Pageable"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Models retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Page"
                }
              }
            }
          },
          "400": {
            "description": "Invalid pagination parameters"
          }
        }
      },
      "post": {
        "tags": [
          "models"
        ],
        "summary": "Create a new model",
        "description": "Create a new model record. Requires authentification.",
        "operationId": "createModel",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ModelDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Model created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ModelDto"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input data"
          },
          "401": {
            "description": "Unauthorized - Invalid or missing JWT token"
          },
          "409": {
            "description": "Model already exists"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
    "/api/v1/manufacturers/{manufacturerName}/models/{modelName}/{year}": {
      "post": {
        "tags": [
          "manufacturers"
        ],
        "summary": "Create car by manufacturer, model for year",
        "description": "Create a new car record for a specific manufacturer, model and year. Requires authentification.",
        "operationId": "createCarByManufacturerModelYear",
        "parameters": [
          {
            "name": "manufacturerName",
            "in": "path",
            "description": "Name of the manufacturer",
            "required": true,
            "schema": {
              "type": "string"
            },
            "example": "Toyota"
          },
          {
            "name": "modelName",
            "in": "path",
            "description": "Name of the model",
            "required": true,
            "schema": {
              "type": "string"
            },
            "example": "Camry"
          },
          {
            "name": "year",
            "in": "path",
            "description": "Year of the car",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            },
            "example": 2025
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "string",
                "description": "ObjectId of the car"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Car created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CarDto"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input data"
          },
          "401": {
            "description": "Unauthorized - Invalid or missing JWT token"
          },
          "404": {
            "description": "Manufacturer or model not found with provided name"
          },
          "409": {
            "description": "Car already exists"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
      "MakeDto": {
        "type": "object",
        "properties": {
          "makeId": {
            "type": "integer",
            "format": "int32"
          },
          "make": {
            "type": "string",
            "maxLength": 255,
            "minLength": 0
          }
        },
        "required": [
          "make"
        ]
      },
      "ModelDto": {
        "type": "object",
        "properties": {
          "modelId": {
            "type": "integer",
            "format": "int32"
          },
          "make": {
            "$ref": "#/components/schemas/MakeDto"
          },
          "model": {
            "type": "string",
            "maxLength": 255,
            "minLength": 0
          }
        },
        "required": [
          "make",
          "model"
        ]
      },
      "CategoryDto": {
        "type": "object",
        "properties": {
          "categoryId": {
            "type": "integer",
            "format": "int32"
          },
          "category": {
            "type": "string",
            "maxLength": 255,
            "minLength": 0
          }
        },
        "required": [
          "category"
        ]
      },
      "CarDto": {
        "type": "object",
        "properties": {
          "objectId": {
            "type": "string",
            "maxLength": 255,
            "minLength": 0
          },
          "model": {
            "$ref": "#/components/schemas/ModelDto"
          },
          "year": {
            "type": "integer",
            "format": "int32"
          },
          "categories": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CategoryDto"
            },
            "uniqueItems": true
          }
        },
        "required": [
          "model",
          "objectId",
          "year"
        ]
      },
...
      "Pageable": {
        "type": "object",
        "properties": {
          "page": {
            "type": "integer",
            "format": "int32",
            "minimum": 0
          },
          "size": {
            "type": "integer",
            "format": "int32",
            "minimum": 1
          },
          "sort": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
      "Page": {
        "type": "object",
        "properties": {
          "totalPages": {
            "type": "integer",
            "format": "int32"
          },
          "totalElements": {
            "type": "integer",
            "format": "int64"
          },
          "first": {
            "type": "boolean"
          },
          "last": {
            "type": "boolean"
          },
          "size": {
            "type": "integer",
            "format": "int32"
          },
          "content": {
            "type": "array",
            "items": {}
          },
          "number": {
            "type": "integer",
            "format": "int32"
          },
          "sort": {
            "$ref": "#/components/schemas/SortObject"
          },
          "pageable": {
            "$ref": "#/components/schemas/PageableObject"
          },
          "numberOfElements": {
            "type": "integer",
            "format": "int32"
          },
          "empty": {
            "type": "boolean"
          }
        }
      },
      "PageableObject": {
        "type": "object",
        "properties": {
          "offset": {
            "type": "integer",
            "format": "int64"
          },
          "sort": {
            "$ref": "#/components/schemas/SortObject"
          },
          "pageNumber": {
            "type": "integer",
            "format": "int32"
          },
          "pageSize": {
            "type": "integer",
            "format": "int32"
          },
          "unpaged": {
            "type": "boolean"
          },
          "paged": {
            "type": "boolean"
          }
        }
      },
      "SortObject": {
        "type": "object",
        "properties": {
          "empty": {
            "type": "boolean"
          },
          "unsorted": {
            "type": "boolean"
          },
          "sorted": {
            "type": "boolean"
          }
        }
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}